+++ scopes/dependencies/dependency-resolver/dependency-resolver.main.runtime.ts
@@ -23,7 +23,6 @@ import {
   CFG_ISOLATED_SCOPE_CAPSULES,
   getCloudDomain,
 } from '@teambit/legacy/dist/constants';
-import { type DependenciesGraph } from '@teambit/legacy/dist/scope/models/version';
 import { ExtensionDataList } from '@teambit/legacy/dist/consumer/config/extension-data';
 import { componentIdToPackageName } from '@teambit/pkg.modules.component-package-name';
 import { DetectorHook } from '@teambit/dependencies';
