+++ scopes/dependencies/pnpm/lockfile-converter.ts
@@ -8,7 +8,7 @@ import {
   type PackageAttributes,
   type DependencyEdge,
   type DependencyNeighbour,
-} from '@teambit/legacy/dist/scope/models/version';
+} from '@teambit/legacy/dist/scope/models/dependencies-graph';
 import { type CalcDepsGraphOptions, type ComponentIdByPkgName } from '@teambit/dependency-resolver';
 import { getLockfileImporterId } from '@pnpm/lockfile.fs';
