+++ scopes/scope/scope/scope.main.runtime.ts
@@ -47,7 +47,8 @@ import { BitId } from '@teambit/legacy-bit-id';
 import { ExtensionDataEntry, ExtensionDataList } from '@teambit/legacy/dist/consumer/config';
 import { EnvsAspect, EnvsMain } from '@teambit/envs';
 import { compact, slice, difference, partition } from 'lodash';
-import { DepEdge, DependenciesGraph } from '@teambit/legacy/dist/scope/models/version';
+import { DependenciesGraph } from '@teambit/legacy/dist/scope/models/dependencies-graph';
+import { DepEdge } from '@teambit/legacy/dist/scope/models/version';
 import { getGlobalConfigPath } from '@teambit/legacy/dist/global-config/config';
 import { invalidateCache } from '@teambit/legacy/dist/api/consumer/lib/global-config';
 import { ComponentNotFound } from './exceptions';
