+++ src/consumer/component/consumer-component.ts
@@ -10,7 +10,8 @@ import { getCloudDomain, BIT_WORKSPACE_TMP_DIRNAME, BuildStatus, DEFAULT_LANGUAG
 import { Doclet, parser as docsParser } from '@teambit/semantics.doc-parser';
 import logger from '../../logger/logger';
 import { ScopeListItem } from '../../scope/models/model-component';
-import Version, { DepEdge, Log, DependenciesGraph } from '../../scope/models/version';
+import { type DependenciesGraph } from '../../scope/models/dependencies-graph';
+import Version, { DepEdge, Log } from '../../scope/models/version';
 import { pathNormalizeToLinux, PathLinux, PathOsBased, PathOsBasedRelative } from '@teambit/toolbox.path.path';
 import { sha1 } from '@teambit/toolbox.crypto.sha1';
 import { ComponentMap } from '@teambit/legacy.bit-map';
